backend/requirements.txt
# FastAPI and core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# Validation and serialization
pydantic==2.5.1
pydantic[email]==2.5.1
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP client for external APIs
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# Environment and configuration
python-dotenv==1.0.0

# Date and time utilities
python-dateutil==2.8.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2  # for testing

# Code quality
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Production server
gunicorn==21.2.0

# Optional: For advanced features (uncomment if needed)
# celery==5.3.4  # For background tasks
# redis==5.0.1   # For caching and task queue
# pillow==10.1.0 # For image processing
# reportlab==4.0.7  # For PDF generation

# WebSocket support (for real-time features)
websockets==12.0

# Logging and monitoring
structlog==23.2.0

# CORS middleware (already included in FastAPI but explicit version)
starlette==0.27.0

# JSON Web Token utilities
pyjwt==2.8.0

# For SQLite fallback in development
aiosqlite==0.19.0

# File handling
aiofiles==23.2.1

# Rich text formatting for development
rich==13.7.0

# Voice and Audio Processing
elevenlabs==1.8.0  # ElevenLabs API client
websocket-client==1.7.0  # Additional WebSocket client support
numpy==1.24.3  # For audio data processing
scipy==1.11.4  # For audio signal processing

# Claude API Integration
anthropic==0.34.0  # Claude API client
tiktoken==0.6.0  # Token counting for Claude prompts
tenacity==8.2.3  # Retry logic for API calls
jinja2==3.1.4  # Template engine for prompt templates

# Learning Intelligence
scikit-learn==1.3.2  # Machine learning for performance analysis
pandas==2.1.4  # Data analysis for learning metrics
textstat==0.7.3  # Text analysis for conversation assessment